refact: Refactorings in tests, made them run individually
